requirements.txt
firebase-admin==6.4.0
pandas>=2.0.0
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
numpy>=1.24.0
